pyproject.toml
[project]
name = "pyinfra-test-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
authors = [{ name = "Chris Mamon", email = "chrisam1993@gmail.com" }]
dependencies = [
  "pyinfra==3.5.0",
]

[project.scripts]
pyinfra-test-example = "pyinfra_test_example:main"

[dependency-groups]
dev = [
  "pytest>=9.0.1",
  "pytest-testinfra>=10.2.2",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
